+++ src/main/java/com/revature/BatchServiceApplication.java
@@ -1,4 +1,4 @@
-package com.revature.batch;
+package com.revature;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
