+++ src/main/java/com/revature/bean/BamUser.java
@@ -1,4 +1,4 @@
-package com.revature.batch.bean;
+package com.revature.bean;
 
 public class BamUser {
