+++ src/main/java/com/revature/bean/Role.java
@@ -1,4 +1,4 @@
-package com.revature.batch.bean;
+package com.revature.bean;
 
 public enum Role {
     NONE, ASSOCIATE, TRAINER, ADMIN;
