+++ src/main/java/com/revature/exception/BatchUpdateException.java
@@ -1,4 +1,4 @@
-package com.revature.batch.exception;
+package com.revature.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -7,20 +7,19 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  * @author Joshua Boudreau, Sonam Sherpa, Marko Miocic (1802-Matt)
  * Last updated: 4/10/18
  */
-@ResponseStatus(value = HttpStatus.NO_CONTENT)
-public class NoBatchException extends RuntimeException{
+@ResponseStatus(value = HttpStatus.BAD_REQUEST)
+public class BatchUpdateException extends RuntimeException {
 
-	private static final long serialVersionUID = 7691014160603213208L;
-	
+	private static final long serialVersionUID = 6173194688911103752L;
+
 	private final String message;
-	
-	public NoBatchException(String message) {
+
+	public BatchUpdateException(String message) {
 		this.message = message;
 	}
-	
+
 	@Override
 	public String getMessage() {
 		return message;
 	}
-
 }
