+++ src/main/java/com/revature/exception/NoBatchException.java
@@ -1,4 +1,4 @@
-package com.revature.batch.exception;
+package com.revature.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -7,19 +7,20 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  * @author Joshua Boudreau, Sonam Sherpa, Marko Miocic (1802-Matt)
  * Last updated: 4/10/18
  */
-@ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class BatchUpdateException extends RuntimeException {
+@ResponseStatus(value = HttpStatus.NO_CONTENT)
+public class NoBatchException extends RuntimeException{
 
-	private static final long serialVersionUID = 6173194688911103752L;
-
+	private static final long serialVersionUID = 7691014160603213208L;
+	
 	private final String message;
-
-	public BatchUpdateException(String message) {
+	
+	public NoBatchException(String message) {
 		this.message = message;
 	}
-
+	
 	@Override
 	public String getMessage() {
 		return message;
 	}
+
 }
