+++ src/main/java/com/revature/logging/JSONify.java
@@ -1,4 +1,4 @@
-package com.revature.batch.logging;
+package com.revature.logging;
 
 
 /**
