+++ src/main/java/com/revature/repository/BatchRepository.java
@@ -1,9 +1,9 @@
-package com.revature.batch.repository;
+package com.revature.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.revature.batch.bean.Batch;
+import com.revature.bean.Batch;
 
 import java.sql.Timestamp;
 import java.util.List;
